requirements.txt
asgiref           3.8.1
blinker           1.9.0
click             8.1.7
colorama          0.4.6
Django            5.1
et_xmlfile        2.0.0
Flask             2.0.1
itsdangerous      2.2.0
Jinja2            3.1.4
lxml              5.3.0
MarkupSafe        3.0.2
numpy             2.1.3
opencv-python     4.10.0.84
openpyxl          3.1.5
packaging         24.2
pandas            2.2.3
pillow            11.0.0
pip               24.2
pytesseract       0.3.13
python-dateutil   2.9.0.post0
python-docx       1.1.2
pytz              2024.2
six               1.16.0
sqlparse          0.5.1
typing_extensions 4.12.2
tzdata            2024.1
Werkzeug          2.0.1